requirements.txt
###### Requirements without Version Specifiers ######
numpy
matplotlib
pillow
opencv-python
